Add view to buy/sell
buy / sell not working yet.
added the functionality in the UI to know if the user is subscribed to option or not, to render UI accordingly.

I figure I'll do it in the UI, although I think maybe and endpoint will be better to know this information, but for the sake of the challenge