Add Buy/Sell Endpoint and Fix holdings endpoint
the buy/sell endpoint could be used sending positives or negatives values to do buy or sell in that order.
This would ideally be done by using PUT and sending the request in the body but my body was always coming empty and I could not get why.

Also fixed the holdings endpoint that was coping the userData object as reference and was updating it adding the name when calling it.